# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * SignageOS - Digital Signage Management Platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}

export interface Client {
  id: number;
  name: string;
  /** Type: gas_station, clinic, other */
  type: string;
  /** @nullable */
  contactName?: string | null;
  /** @nullable */
  contactPhone?: string | null;
  /** @nullable */
  address?: string | null;
  active?: boolean;
  screenCount?: number;
  createdAt: string;
}

export interface ClientInput {
  /** @minLength 1 */
  name: string;
  type: string;
  contactName?: string;
  contactPhone?: string;
  address?: string;
}

export interface ClientUpdate {
  name?: string;
  type?: string;
  contactName?: string;
  contactPhone?: string;
  address?: string;
  active?: boolean;
}

export interface Screen {
  id: number;
  name: string;
  clientId: number;
  /** @nullable */
  clientName?: string | null;
  /** Unique code used by the TV player URL */
  code: string;
  /** @nullable */
  location?: string | null;
  /** online, offline, unknown */
  status: string;
  /** @nullable */
  lastSeen?: string | null;
  /** @nullable */
  activePlaylistName?: string | null;
  createdAt: string;
}

export interface ScreenInput {
  /** @minLength 1 */
  name: string;
  clientId?: number;
  location?: string;
}

export interface PairScreenInput {
  /** @minLength 1 */
  pairingCode: string;
  /** @minLength 1 */
  name: string;
  location?: string;
}

export interface PairScreenResult {
  id: number;
  name: string;
  code: string;
  /** @nullable */
  location?: string | null;
  status: string;
  createdAt: string;
}

export interface BroadcastInput {
  playlistId: number;
}

export interface BroadcastResult {
  count: number;
}

export interface ScreenUpdate {
  name?: string;
  location?: string;
  clientId?: number;
}

export interface MediaFile {
  id: number;
  name: string;
  /** image, video */
  type: string;
  url: string;
  /** @nullable */
  thumbnailUrl?: string | null;
  /** @nullable */
  durationSeconds?: number | null;
  /** @nullable */
  clientId?: number | null;
  createdAt: string;
}

export interface MediaInput {
  /** @minLength 1 */
  name: string;
  type: string;
  url: string;
  thumbnailUrl?: string;
  durationSeconds?: number;
  clientId?: number;
}

export interface Playlist {
  id: number;
  name: string;
  /** @nullable */
  clientId: number | null;
  /** @nullable */
  clientName?: string | null;
  itemCount: number;
  totalDurationSeconds?: number;
  createdAt: string;
}

export interface PlaylistItem {
  id: number;
  playlistId: number;
  mediaId: number;
  /** @nullable */
  mediaName?: string | null;
  /** @nullable */
  mediaUrl?: string | null;
  /** @nullable */
  mediaType?: string | null;
  position: number;
  durationSeconds: number;
}

export interface PlaylistDetail {
  id: number;
  name: string;
  /** @nullable */
  clientId?: number | null;
  items: PlaylistItem[];
}

export interface PlaylistInput {
  /** @minLength 1 */
  name: string;
  clientId?: number;
}

export interface PlaylistUpdate {
  name?: string;
  clientId?: number;
}

export interface PlaylistItemInput {
  mediaId: number;
  durationSeconds: number;
  position?: number;
}

export interface Schedule {
  id: number;
  screenId: number;
  /** @nullable */
  screenName?: string | null;
  playlistId: number;
  /** @nullable */
  playlistName?: string | null;
  /** @nullable */
  startTime?: string | null;
  /** @nullable */
  endTime?: string | null;
  /**
     * Comma-separated day numbers: 0=Sun, 1=Mon, ... 6=Sat
     * @nullable
     */
  daysOfWeek?: string | null;
  active?: boolean;
  createdAt: string;
}

export interface ScheduleInput {
  screenId: number;
  playlistId: number;
  startTime?: string;
  endTime?: string;
  daysOfWeek?: string;
  active?: boolean;
}

export interface ScheduleUpdate {
  playlistId?: number;
  startTime?: string;
  endTime?: string;
  daysOfWeek?: string;
  active?: boolean;
}

export interface CountByType {
  type: string;
  count: number;
}

export interface DashboardStats {
  totalClients: number;
  totalScreens: number;
  onlineScreens: number;
  totalPlaylists: number;
  totalMedia: number;
  clientsByType?: CountByType[];
}

export interface ActivityItem {
  id: number;
  action: string;
  entityType: string;
  entityName: string;
  createdAt: string;
}

export interface PlayerItem {
  mediaUrl: string;
  mediaType: string;
  durationSeconds: number;
  mediaName?: string;
}

export interface PlayerPayload {
  screenId: number;
  screenName: string;
  items: PlayerItem[];
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
  /** @nullable */
  pairingCode?: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export interface LogoutSuccess {
  success: boolean;
}

export interface ErrorEnvelope {
  error: string;
}

export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
returnTo?: string;
};

export type ListScreensParams = {
clientId?: number;
};

export type ListMediaParams = {
clientId?: number;
};

export type ListPlaylistsParams = {
clientId?: number;
};

export type ListSchedulesParams = {
screenId?: number;
};