# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SignageOS - Digital Signage Management Platform
servers:
  - url: /api
    description: Base API path
...
paths:
  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the OIDC authorization endpoint.
...
  /mobile-auth/token-exchange:
    post:
      tags: [Auth]
      operationId: exchangeMobileAuthorizationCode
      summary: Exchange a mobile OIDC code for a session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileTokenExchangeRequest'
      responses:
        '200':
          description: Session created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileTokenExchangeSuccess'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Token did not contain usable claims.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Token exchange failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /mobile-auth/logout:
    post:
      tags: [Auth]
      operationId: logoutMobileSession
      summary: Delete a mobile session token
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
      responses:
        '200':
          description: Logout processed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'
...
  /media:
    get:
      operationId: listMedia
      tags: [media]
      summary: List all media files
      parameters:
        - name: clientId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of media files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MediaFile"
    post:
      operationId: createMedia
      tags: [media]
      summary: Add a media file (URL-based)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MediaInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaFile"
  /media/{id}:
    get:
      operationId: getMedia
      tags: [media]
      summary: Get a media file by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Media file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaFile"
    delete:
      operationId: deleteMedia
      tags: [media]
      summary: Delete a media file
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /playlists:
    get:
      operationId: listPlaylists
      tags: [playlists]
      summary: List all playlists
      parameters:
        - name: clientId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of playlists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Playlist"
    post:
      operationId: createPlaylist
      tags: [playlists]
      summary: Create a new playlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaylistInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Playlist"
  /playlists/{id}:
    get:
      operationId: getPlaylist
      tags: [playlists]
      summary: Get a playlist with its items
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Playlist with items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaylistDetail"
    patch:
      operationId: updatePlaylist
      tags: [playlists]
      summary: Update a playlist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaylistUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Playlist"
    delete:
      operationId: deletePlaylist
      tags: [playlists]
      summary: Delete a playlist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /playlists/{id}/items:
    post:
      operationId: addPlaylistItem
      tags: [playlists]
      summary: Add a media item to a playlist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaylistItemInput"
      responses:
        "201":
          description: Item added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaylistItem"
  /playlists/{id}/items/{itemId}:
    delete:
      operationId: removePlaylistItem
      tags: [playlists]
      summary: Remove a media item from a playlist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
...
components:
  schemas:
...
    MediaFile:
      type: object
      required: [id, name, type, url, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          description: "image, video"
        url:
          type: string
        thumbnailUrl:
          type: ["string", "null"]
        durationSeconds:
          type: ["integer", "null"]
        clientId:
          type: ["integer", "null"]
        createdAt:
          type: string
    MediaInput:
      type: object
      required: [name, type, url]
      properties:
        name:
          type: string
          minLength: 1
        type:
          type: string
        url:
          type: string
        thumbnailUrl:
          type: string
        durationSeconds:
          type: integer
        clientId:
          type: integer
    Playlist:
      type: object
      required: [id, name, clientId, itemCount, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        clientId:
          type: ["integer", "null"]
        clientName:
          type: ["string", "null"]
        itemCount:
          type: integer
        totalDurationSeconds:
          type: integer
        createdAt:
          type: string
    PlaylistDetail:
      type: object
      required: [id, name, items]
      properties:
        id:
          type: integer
        name:
          type: string
        clientId:
          type: ["integer", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/PlaylistItem"
    PlaylistItem:
      type: object
      required: [id, playlistId, mediaId, position, durationSeconds]
      properties:
        id:
          type: integer
        playlistId:
          type: integer
        mediaId:
          type: integer
        mediaName:
          type: ["string", "null"]
        mediaUrl:
          type: ["string", "null"]
        mediaType:
          type: ["string", "null"]
        position:
          type: integer
        durationSeconds:
          type: integer
    PlaylistInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        clientId:
          type: integer
    PlaylistUpdate:
      type: object
      properties:
        name:
          type: string
        clientId:
          type: integer
    PlaylistItemInput:
      type: object
      required: [mediaId, durationSeconds]
      properties:
        mediaId:
          type: integer
        durationSeconds:
          type: integer
        position:
          type: integer
...
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
        pairingCode:
          type: [string, 'null']
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'
    MobileTokenExchangeRequest:
      type: object
      required: [code, code_verifier, redirect_uri, state]
      properties:
        code:
          type: string
          minLength: 1
        code_verifier:
          type: string
          minLength: 1
        redirect_uri:
          type: string
          minLength: 1
        state:
          type: string
          minLength: 1
        nonce:
          type: string
          minLength: 1
    MobileTokenExchangeSuccess:
      type: object
      required: [token]
      properties:
        token:
          type: string
    LogoutSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      schema:
        type: string