# lib/api-zod/src/generated/types/authUser.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * SignageOS - Digital Signage Management Platform
 * OpenAPI spec version: 0.1.0
 */

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
  /** @nullable */
  pairingCode?: string | null;
}